backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1

# WebSocket
eventlet==0.34.2
python-socketio==5.10.0

# Security
bcrypt==4.1.1
pyotp==2.9.0
qrcode==7.4.2
PyJWT==2.8.0

# Crypto/Blockchain
bitcoinlib==0.6.14
eth-account==0.10.0
web3==6.12.0
tronpy==0.4.0
stellar-sdk==9.0.0
solana==0.31.0
base58==2.1.1
mnemonic==0.20
hdwallets==0.1.2
ecdsa==0.18.0

# Validation
marshmallow==3.20.1
email-validator==2.1.0
phonenumbers==8.13.26

# HTTP client
requests==2.31.0
aiohttp==3.9.1

# Task queue (optional, for background jobs)
celery==5.3.4

# Utils
python-dotenv==1.0.0
python-dateutil==2.8.2
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production server
gunicorn==21.2.0
